vertex_agent/bq_tool_handler/requirements.txt
flask==3.*
google-cloud-bigquery==3.*
gunicorn==21.*
